Add hover functionality for nav tabs
This commit is very comprehensive because of changes made to existing
code.

Location-tab and order-tab ids were converted to classes. Background
color and color transition was added to these classes.

Hovers were not working because nav was hidden under other content.
Z index was added to nav to fix this issue.

Event listener for mouseenter/mouseleave was added to the for loop for
the last two tabs in the nav.

this.classList.toggle was added to changeColor function to add the
hover background color and font color to the last two tabs.